New Storage Types  - Added the ability for ODS to now use memory only compared to needed a file.  - Changed the Compressor methods.  - Split code up into internal classes.  - Added the export method.
