+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>me.ryandw11</groupId>
     <artifactId>ods</artifactId>
-    <version>1.0.3</version>
+    <version>NewLayout-SNAPSHOT</version>
 
     <build>
         <sourceDirectory>src/main/java</sourceDirectory>
