+++ src/main/java/me/ryandw11/ods/compression/Compressor.java
@@ -15,18 +15,18 @@ public interface Compressor {
     /**
      * Get the input stream for the compression.
      *
-     * @param file The file input stream for the file.
+     * @param inputStream The file input stream for the file.
      * @return The compression input stream.
      * @throws IOException An IOException if one occurs.
      */
-    InputStream getInputStream(FileInputStream file) throws IOException;
+    InputStream getInputStream(InputStream inputStream) throws IOException;
 
     /**
      * Get the output stream from the compression.
      *
-     * @param file The file output stream for the file.
+     * @param outputStream The file output stream for the file.
      * @return The compression output stream.
      * @throws IOException An IOException if one occurs.
      */
-    OutputStream getOutputStream(FileOutputStream file) throws IOException;
+    OutputStream getOutputStream(OutputStream outputStream) throws IOException;
 }
