+++ src/main/java/me/ryandw11/ods/internal/ODSInternal.java
@@ -1,0 +1,34 @@
+package me.ryandw11.ods.internal;
+
+import me.ryandw11.ods.Tag;
+import me.ryandw11.ods.compression.Compressor;
+
+import java.util.List;
+
+/**
+ * This interface is for the different ways data can be stored / read. viz. File and Memory
+ * <p>
+ * To see the method explanations please view the individual implementation: {@link ODSFile}, {@link ODSMem}.
+ * You can also view the main explanations at {@link me.ryandw11.ods.ObjectDataStructure}.
+ */
+public interface ODSInternal {
+    <T> T get(String key);
+
+    List<Tag<?>> getAll();
+
+    void save(List<? extends Tag<?>> tags);
+
+    void append(Tag<?> tag);
+
+    void appendAll(List<Tag<?>> tags);
+
+    boolean find(String key);
+
+    boolean delete(String key);
+
+    boolean replaceData(String key, Tag<?> replacement);
+
+    void set(String key, Tag<?> value);
+
+    byte[] export(Compressor compressor);
+}
